Module 3
Data Insights and Presentation:
Using visualisations to present insights.

 After building the model, the results need to be presented to the client. 
 Visualisations such as interactive dashboards often help us highlight key findings and convey our ideas in a more succinct manner.So,there is a need to support our results with the use of visualisations. 

Develop a dashboard that can be presented to the client at our next meeting. Displaying the data summary and results of the analysis in a dashboard (see tools/references for assistance).  Maximum of 3 dashboard views/tabs, creativity in layout and presentation is welcome.  

It is important to keep in mind the business context when presenting your findings:
1. What are the trends in the underlying data?
2. Which customer segment has the highest customer value?
3. What do you propose should be Sprocket Central Pty Ltd ’s marketing and growth strategy?
4. What additional external datasets may be useful to obtain greater insights into customer preferences and propensity to purchase the products?
Specifically, your presentation should specify who Sprocket Central Pty Ltd’s marketing team should be targeting out of the new 1000 customer list as well as the broader market segment to reach out to.