Module 1
Task 1
Data Quality Assessment :
Assessment of data quality and completeness in preparation for analysis.

The importance of optimising the quality of customer datasets cannot be underestimated. The better the quality of the dataset, the better chance you will be able to use it drive company growth.
Assessing the quality of the data and providing ways to clean the data and improve overall data of dataset for further analysis.

The client provided KPMG with 3 datasets:

1. Customer Demographic
2. Customer Addresses
3. Transactions data in the past 3 months

The client has asked our team to assess the quality of their data; as well as make recommendations on ways to clean the underlying data and mitigate these issues.So, decided to start the preliminary data exploration and identify ways to improve the quality of Sprocket Central Pty Ltd’s data.